requirements.txt
Flask==1.0.2
gunicorn==19.9.0
Flask-SQLAlchemy==2.3.2
Flask-Dance==1.1.0
pendulum==2.0.3
sqlalchemy_utils==0.33.5
Flask-Login==0.4.1
blinker==1.4
pyOpenSSL==18.0.0
Flask-Cors==3.0.6
psycopg2-binary
